moving process.Manager to process/globals.go
